pyproject.toml
[project]
name = "lakehouse-agent"
version = "0.1.0"
description = "Autonomous Lakehouse Operations Agent for Databricks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "databricks-sdk>=0.102.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "fastmcp>=3.2.4",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.1.0",
    "langfuse>=4.2.0",
    "langgraph>=1.1.6",
    "uvicorn>=0.44.0",
]
